Practice Exercise
Practice exercise scenario: "Qanva is a startup that helps users manage their time more effectively and stay on track towards their daily goals. They just received seed funding and will hire their first designer (you). The first iteration of the product was designed by the founder and the VP of Engineering. As you can see in the provided screenshot (Col_screen2), the app exhibits a poor use of color. As the first designer in the company, look for opportunities to improve this design. Generate a new color palette, pick a vibrant color palette that benefits clarity, and the different types of tasks scheduled are distinct."