Practice Exercise
Practice Exercise: "Changelly is a secure online platform for buying and selling cryptocurrency. This landing page was built for conversion purposes, getting more new, and the effectiveness hugely relies on your design decisions. The business goal for this page is to get more users on the platform, and the primary metric is conversion, which is to get users to create an account. Look at the provided screenshot below, which exhibits poor use of positive and negative space. Look for opportunities to improve this design."